converter/requirements.txt
Django==5.0.4
django-payments==2.0.0
django-payments-payu==1.4.0
django-phonenumber-field==7.3.0
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
pillow==10.3.0
PyJWT==2.8.0
PyYAML==6.0.1
Jinja2==3.1.4
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
simplejson==3.19.2
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1

